Move _relevant from cli.py to storage.py
